JS: React 5
This code creates a Progress component that displays a progress bar. Here's the step-by-step breakdown:
Import React: At the beginning of the file, we import the React library.
Progress Class: We define the Progress component class, inheriting from React.Component.
render() Method: This method is required in a React component. It handles the rendering of the component.
Props Destructuring: We destructure the percentage property from the passed props. The progress percentage to be displayed will be taken from this property.
divStyle Object: We create a divStyle object that will hold the style for the progress bar. The width property in this object is set to ${percentage}% to set the width of the bar in percentages.
ariaValuenow Variable: We create an ariaValuenow variable and assign it the value of percentage.
JSX: We use JSX to render the component. Inside the <div> with the class progress, we place a <div> with the class progress-bar that represents the progress bar. We use the aria-valuenow, aria-valuemin, and aria-valuemax values for accessibility. The progress bar's style is set through the style property using the divStyle object.
Closing the Component: We close the component.
This way, the Progress component renders a progress bar with dynamic styling based on the provided percentage prop.