JS: React 9
JS: React - Events:
Events in React represent actions that users or other sources can perform on interface elements. React handles events similarly to handling events in regular JavaScript, but with some differences and optimizations.
In this code, a Carousel component is created, representing an image carousel. Let's see how events are handled in this code:
setNext and setPrev methods: These methods are called when the "Next" and "Previous" buttons are clicked, respectively. They modify the currentIdx state to switch the display between images.
render() method: In the render() method, the "Next" and "Previous" buttons are assigned event handlers using onClick. When the button is clicked, the corresponding method (setNext or setPrev) is called, leading to a change in the component's state and a re-render.
onClick event: The onClick attribute in JSX defines what code should be executed when an element (in this case, a button) receives a click. In this code, onClick is used to call the setNext and setPrev methods.