JS: React 10
JS: React - Automata-driven Programming:
Automata-driven programming is an approach in programming where the behavior of a program is described using finite state machines. This approach allows describing and managing states and transitions between them. In React, automata-driven programming can be used to create components that can be in different states and respond to events by transitioning between states.
In this code, a Collapse component is created, representing a collapsible and expandable block. Let's see how automata-driven programming is applied in this code:
Constructor constructor(props): In the component's constructor, the initial opened state is set based on the props passed. The opened value will store the collapsed/expanded state of the block.
Method handleToggle(): This method is called when the "Link with href" is clicked. It uses functional state update with setState() to change the opened state to its opposite value.
Method render(): In this method, the opened state value and the text and opened props are used to dynamically generate classes and attributes. Depending on the value of opened, the classes and aria-expanded attributes are adjusted to match the current state.