JS: React 11
JS: React - Forms:
Forms in React allow users to input and submit data to the server. In this example, a MyForm component is created, representing a form with various input fields. Let's explore the key aspects of this form:
Constructor constructor(props): In the constructor, the initial state of the form is set, including field values and the current submission state (submittingState). By default, the form is in the filling state.
Method handleChangeField(): This method is called when the values of input fields change. It uses a dynamically computed key of the state object form to update the value of the field based on its name.
Methods handleBackToForm() and handleSubmitForm(): These methods are used to manage the form's state. The first returns the form to the filling state after submission, and the second performs form submission.
Method renderForm(): This method generates the markup for the form. It uses the values from the form state to set the values of input fields and properties of form elements.
Method renderResult(): This method generates the markup displaying the results of the form submission. It displays all fields and their values.
Method render(): In this method, the submittingState is used to determine which markup to display. If the form is in the filling state, the renderForm() method is displayed; if the form is submitted, the renderResult() method is displayed.