JS: React 12
JS: React - Immutability:
Immutability is a concept in React where data in the application is not directly modified. Instead, a new copy of the data is created when it changes. This helps avoid unpredictable behavior and simplifies tracking changes.
In this example, a Component is created to demonstrate the concept of immutability when working with the items state array. Let's look at the key points:
Constructor constructor(props): In the constructor, the initial state is set, where the items array is initialized as empty.
Method handleCount(): This method modifies the value in the items array by adding or subtracting the value. The get() function from the Lodash library is used for safe access to the value in the array.
Methods handleInc() and handleDec(): These methods call handleCount() with the appropriate argument to increase or decrease the value.
Method handleRemove(): This method removes an element from the items array by filtering elements based on their id.
Method renderLog(): This method renders the elements of the items array as buttons that can be removed.
Method render(): This method renders the component interface, including buttons for increasing and decreasing the value, and a list of buttons from the renderLog() method.