JS: React 13
JS: React - Nested Components:
Nested components in React refer to the practice of breaking down the interface into smaller and reusable components. This helps break down a complex interface into simpler and more manageable parts.
In this example, two components are created: Item and TodoBox.
Component Item: This component displays an individual task. It receives the task (task object) and onRemove (function for task removal). Inside the component, a button for removing the task and the task's text are displayed.
Component TodoBox: This component represents a list of tasks. It holds the state for the new task text and an array of tasks. It includes methods for adding and removing tasks.
handleChangeTask(): Updates the newTaskText state when the input field value changes.
handleRemoveTask(): Removes a task from the array based on the provided id.
handleSubmitForm(): Adds a new task to the tasks array and resets the new task text.
renderForm(): Renders the form for adding a new task.
render(): Displays the form and the list of tasks using the Item component. Each task is passed to the Item component as task, along with the handleRemoveTask function for task removal.