JS: React 17
This code demonstrates an example of asynchronous data handling using the Axios library in React:
Autocomplete Component:
This component implements an autocomplete text field for searching countries.
In the component's constructor, initial states countries (list of countries) and text (search text) are set.
handleChangeText:
This is an asynchronous method that is called when the text in the input field changes.
If the input field is cleared, the text and countries states are reset.
If there is text in the input field, it's saved in the text state, and then an asynchronous request to the server is made using Axios.
The result of the request (list of countries) is saved in the countries state.
renderCountries:
This method renders the list of countries as an unordered list.
render:
The render method displays the component in the web interface.
It includes a form with an input field for entering text.
If there are search results (list of countries), they are also displayed.
In summary, the Autocomplete component allows the user to search for countries based on the entered text. When the text changes, an asynchronous request is made to the server to retrieve a list of countries that match the entered text, and this list is displayed to the user.