JS: React 18
This code demonstrates the use of React component's lifecycle in the context of a TODO list:
TodoBox Component:
In the constructor of the component, initial states are initialized: newTaskText (text of a new task) and tasks (list of tasks).
componentDidMount(): a lifecycle method that is called after the component has been inserted into the DOM. In this case, it's used to fetch tasks from the server when the page is initially loaded.
handleChangeText:
This method is an event handler for changes in the input field for the new task. It updates the newTaskText state.
handleFinishTask and handleActivateTask:
Asynchronous event handler methods for finishing and activating a task, respectively. They send PATCH requests to the server to update the task's status.
After successfully updating the status, these methods update the list of tasks in the component's state using the immutability-helper library to update state without mutation.
handleSubmitForm:
An asynchronous event handler method for submitting the form to add a new task. It sends a POST request to the server to create a new task.
After successfully creating the task, this method updates the newTaskText state and the list of tasks.
fetchTasks:
An asynchronous method for fetching the list of tasks from the server. It performs a GET request and updates the tasks state.
renderFinishedTasks and renderActiveTasks:
Methods for rendering finished and active tasks, respectively. They create Item components for each task.
renderForm:
A method for rendering the form to add a new task.
render:
The method for rendering the main content of the component. It invokes the task rendering methods and the form rendering method.
Item Component:
This component is responsible for displaying an individual task in the list.
The onClick method is set depending on the task's status, allowing tasks to be finished or activated.
The task link is displayed, and if the task is finished, it is crossed out using the <s> tag.
This example demonstrates how to use React component's lifecycle for asynchronous data fetching from a server and how to handle user actions to change the component's state without directly mutating the data.