relational databases 3
Data is retrieved in sql using the SELECT command, example: SELECT username, email FROM users; LIMIT - limit on the number of records that are selected from the database. SQL does not guarantee order in the selections above. If we do not specify sorting, then we cannot know in what order the data will be returned. Therefore, in such selections there is always an ORDER BY section; if you need to sort in reverse order, you need to add the DESC keyword. For example, SELECT * FROM users ORDER BY birthday DESC.