relational databases 5
Normal form is a requirement that minimizes data redundancy that can potentially lead to logical errors. There are six normal forms in total, which include certain requirements. With each subsequent level, the requirements become more stringent, as they include the previous levels. First normal form comes down to three rules:
-Each table cell can store only one value
-All data in one column can only be of one type
-Each entry in the table must be uniquely different from other entries. PRIMARY KEY - A field or set of fields that contains a unique value for each record. The primary key cannot change; its value uniquely identifies any record in the table.
Let's look at two types of primary key:
Natural - when values ​​from the outside world are used, for example, email, full name or passport data. In this case, you need to make sure that the key will not be repeated. Such primary keys are rarely used due to their unreliability. They are often not unique and may change or be repeated. For example, the passport number changes when the document is changed
Surrogate - when automatically generated unique values ​​are used. This key is supported by any database out of the box. Sometimes these are just numbers, and sometimes they are complex number-letter strings or hashes.