relational databases 8
Auto-generation of primary key. The primary key in databases is usually filled in automatically using the capabilities built into the database. Automatic generation of a primary key is often called autoincrement. Which translates to automatic increment and is reminiscent of the increment operation from Programming++. This auto-increment is known as GENERATED AS IDENTITY. Here is its structure from the documentation:
column_name type GENERATED { ALWAYS | BY DEFAULT } AS IDENTITY[ ( sequence_option ) ]
- Data type can be SMALLINT, INT or BIGINT
- GENERATED ALWAYS - will not allow you to add a value yourself using UPDATE or INSERT
- GENERATED BY DEFAULT - unlike the previous option, this option allows you to add values ​​yourself.