relational databases 9
In this lesson we will learn what an ontology is and how it helps to describe the subject area of ​​a program and show the relationships of its entities. As a result, we will analyze each type of connection and understand in what cases they will be useful. Let's also look at how these connections are visualized. Let's take a look at Hexlet since you're familiar with it. To understand the subject area of ​​Hexlet, you need to identify key concepts - entities around which all logic is built. For educational resources, these are usually “course” and “lesson”. But in reality there are much more entities. But entities do not exist on their own, but are in relationship with each other.
These connections have specific names:
One to many, one-to-many or o2m. For example, when one lesson can only be in one course, but the course contains many lessons
One to one, one-to-one or o2o. For example, on Hexlet such a connection is established between a user and a Facebook account
Many to many, many-to-many or m2m. For example, when many users can take one course, and one user can take many courses. The description of the objects of the domain under consideration and the connections between them is called the domain ontology.