relational databases 10
We continue to talk about relational databases, the topic of Constraints, in this lesson we examined such constraints as: The primary key uniquely identifies each record within the table. Specified using the PRIMARY KEY clause, which is added after the type when the table is created. There can only be one primary key in a table, and a surrogate key is used for it - an identifier that has no physical meaning.
A foreign key is a constraint that links a specified column to data from another table. This restriction guarantees data integrity. That is, attempting to delete data that has dependent data will result in an error. * If you try to insert a record into orders with a product_id value that is not in the products table, an error will occur.
* If you try to delete a record from products that is referenced from the orders table, an error will occur.
Not Null - This constraint indicates that the column cannot contain null values. It can be combined with other constraints, such as foreign keys. The primary key is NOT NULL by default.
To ensure that the values ​​of a single field are unique, the UNIQUE keyword is used. It is usually used in the description of a specific field when a table is created. UNIQUE is never used in conjunction with a primary key because it is unique by definition.