relational databases 12
In this tutorial, we learned how to sort data in a database using the ORDER BY query. Key facts: 1. The order of data in a database selection is not defined by default and can change. 2. The ORDER BY query allows you to sort data by a specific field or multiple fields. 3. Sorting can be done in forward (ASC) or reverse (DESC) order. 4. When sorting by several fields, the sort order is set for each field separately. 5. When sorting NULL values, they can be located at the beginning or end of the selection, depending on the specified NULLS FIRST or NULLS LAST parameter. Summary: The ORDER BY query allows you to order data in a database by a specific field or multiple fields in forward or reverse order. You can also control the sorting of NULL values.