relational databases 13
In this tutorial, we learned how to use conditions to select a portion of data from a database. We learned how to use comparison operators (=, !=, >, <, >=, <=) to select data, and how to use logical operators (AND, OR) to combine conditions. We also looked at the special LIKE and IN operators for partial match search and value set search, respectively. So, the main points that we learned in this lesson: - WHERE is used to specify the conditions for retrieving data from the database. - Comparison operators (=, !=, >, <, >=, <=) are used to compare field values. - Logical operators (AND, OR) are used to combine conditions. - The LIKE and IN operators are used for partial match search and value set search respectively. Now you know how to use conditions to retrieve the desired data from a database, and you can apply this knowledge in your work with databases.