relational databases 14
In this lesson, we learned the LIMIT function, which allows you to limit the selection of data from a table. We learned that LIMIT is used to output only a certain number of rows from a table, for example 10 records. To guarantee a specific sort order, we must use ORDER BY together with LIMIT. We also looked at how to implement pagination so that the user can view the following pages of data. To do this, we used OFFSET, which allows you to skip a certain number of lines before outputting data. As a result, using LIMIT, ORDER BY and OFFSET, we can limit the data sample, set sorting and view data for all samples.