relational databases 15
A thought lesson on how to use the DISTINCT function in SQL to find unique records in a table. DISTINCT Ensures that duplicate rows occur in the result set. This tutorial shows how to use DISTINCT to get unique usernames from the users table. Examples of using DIFFERENT with multiple poles and combined with aggregate functions are also shown. At the end of the lesson there is also an example of how to use DISTINCT ON to select individual records based on specific fields. Lesson summary: DISTINCT Allows you to find records in a table and check for duplicate data.