relational databases 16
In this lesson, we looked at aggregate functions in SQL. Aggregate functions allow you to perform operations on data groups, such as counting the number of records, finding the maximum or minimum value, summing values and finding the arithmetic mean.

The most popular aggregate functions in SQL:

COUNT(): Counts the number of records in the selection.
MAX(): finds the maximum value.
MIN(): finds the minimum value.
SUM(): finds the sum of the values.
AVG(): Finds the arithmetic mean of the values.
These functions can be very useful when analyzing data and performing various operations with it. They allow you to reduce the amount of code and simplify working with data.

It is important to remember that aggregate functions operate on groups of data, so they are often used with GROUP BY operators or in combination with other functions and conditions.

Conclusions:

Aggregate functions in SQL allow you to perform operations on groups of data.
The most popular aggregate functions in SQL are COUNT(), MAX(), MIN(), SUM(), AVG().
Aggregate functions simplify working with data and reduce the amount of code.