relational databases 17
The lesson explains how the GROUP BY function works in SQL. The GROUP BY function allows you to group rows by a specific attribute and perform aggregate functions within each group. For example, you can output the total number of topics for each user.

Request example: SELECT user_id, COUNT(*) FROM topics GROUP BY user_id;

The result will contain the number of topics for each user.

The lesson also discusses the use of sorting and limiting results, as well as the use of aliases for aggregate values and other fields in the query.

Next, the lesson explains why you cannot use fields that are not specified in GROUP BY or aggregate functions. This is because grouping creates independent groups of records, and the database cannot uniquely select a value for such fields.

The lesson also discusses the use of HAVING to select groups based on aggregate.

Lesson summary: the GROUP BY function allows you to group rows by a specific attribute and perform aggregate functions within each group. It is useful for analyzing data and performing complex tasks in a single query.