relational databases 19
In this lesson, we learned how to execute queries inside a transaction. A transaction is a block of requests that is framed by BEGIN and COMMIT requests. Multiple queries can be executed within a transaction, and if all of them are completed successfully, the changes are saved to the database. If an error occurs, all changes are rolled back.
We also reviewed the requirements for a transactional system that guarantee its reliability. These requirements are known as ACID:
Atomicity: The transaction must be either fully executed or not executed at all. There are no intermediate states.
Consistency: A completed transaction must maintain database consistency. The data must be valid and comply with certain rules.
Isolation: Parallel transactions should not affect each other. Each transaction must work in an isolated space.
Durability: Changes made by a successfully completed transaction must be saved even in the event of system failures.
ACID requirements ensure the reliability and integrity of the data in the database, ensuring that operations are performed safely and predictably.