Ansible 1
This lesson discusses the use of Ansible, a console utility for configuring servers and deploying applications. Ansible allows you to automate various operations on remote servers, such as installing and updating packages, preparing the machine for application code, and other routine tasks.
The lesson discusses the Ansible - ad-hoc mode of operation, when requests to servers are executed directly from the command line without creating additional files. To make a request, you must specify the name of the server or its IP address, as well as user data for SSH connection.

The command examples in the lesson demonstrate checking the availability of the server by IP address and executing the ping command on a remote machine. The use of an inventory file for convenient infrastructure management is also considered.
Lesson summary: Ansible is a powerful utility for automating server configuration and application deployment. It allows you to make requests on remote servers, simplifying routine operations and infrastructure management.