Ansible 3 Playbook
In this lesson, we learned the basic concepts in Ansible, including playbooks, tasks, and modules. A playbook is a yaml file that describes the tasks that need to be performed on certain servers. Tasks consist of modules that define commands that are executed on servers. The playbook example shows the installation and removal of the redis-server package on a group of webservers servers. Each task has a name and a list of parameters for the module. The example uses the apt module to work with packages in Ubuntu. To execute the playbook, the ansible-playbook command is used, which passes the path to the playbook file and the inventory file. When executing a playbook, an error may occur due to insufficient permissions. In this case, you can use the user switching mechanism using the become parameter. By default, it uses sudo to switch to root. After adding the become: yes parameter to the tasks, the playbook is running successfully.