Ansible 4 Tags
This lesson examines the use of tags in Ansible to selectively perform tasks. Tags allow you to optimize the process of making a playbook, especially when working with large playbooks.
Using tags allows you to select and run only the necessary tasks. Tasks can be tagged with multiple tags, which makes them easier to classify and selectively complete.
The example shows the use of tags for tasks related to the installation and configuration of Nginx. The tasks are tagged with "nginx" and "config". When starting a playbook with the -t option and specifying a tag, only tasks with this tag will be completed.
The lesson also discusses the --strip-tags option, which allows you to exclude tasks with certain tags.
Lesson summary: using tags in Ansible allows you to selectively perform tasks, which makes it easier to debug and optimize the playback process.