Ansible 5 Handlers
In this lesson, we studied the handlers mechanism in Ansible. Handlers allow us to associate tasks that can generate changes with certain actions that must be performed after these changes. In our example, we used handlers to restart the nginx service after updating its configuration file.
So, a brief summary of the lesson:
Handlers in Ansible allow us to define tasks that must be completed after changes caused by other tasks.
Handlers communicate with tasks using the notify key.
Handlers are executed only if the task that caused the changes was completed successfully.
Handlers are executed only after the completion of the entire playbook.
Using handlers allows us to manage the order of tasks more flexibly and avoid unnecessary restarts of services.