Ansible 6 Variables
This lesson discusses the use of variables in Ansible. Variables allow you to describe configuration parameters in one place and reuse them multiple times.
The example shows the use of variables to determine the paths to files that need to be updated on remote servers. Instead of duplicating paths, the root_dir variable is used, which contains the common path /var/tmp/www/.
Ansible uses the Jinja 2 template engine to substitute variables. Before processing the YAML file, Ansible runs it through a template engine and replaces the {{variable name}} constructs with the values of these variables. It is important to use quotation marks for variables so that YAML is recognized correctly.
The lesson also explains the difference between the template and copy tasks. The template task allows you to run files through Jinja 2 and perform variable substitutions in them. This is convenient for working with configuration files.
Lesson summary: variables in Ansible allow you to describe configuration parameters in one place and reuse them repeatedly. They are used using the Jinja 2 template engine and allow you to perform value substitutions in files.