Ansible 8 Facts
This example shows how to use facts in Ansible to get information about the target machine and use this information in a configuration file template.
At the beginning of the playbook, the host on which the task will be performed is indicated.
The "update postgresql config" task uses the ansible module.builtin.template to copy the templates/postgresql.conf template.j2 to the target machine in the path /var/tmp/postgresql.conf.
In the templates/postgresql.conf template.j2 uses the ansible_facts variable['memtotal_mb'] to determine the buffer size. In this case, the buffer will use 10% of the total memory of the target machine.
Sample code:
- hosts: dbservers
  tasks:
    - name: update postgresql config
      ansible.builtin.template:
        src: templates/postgresql.conf.j2
        dest: /var/tmp/postgresql.conf
An example of the content of the templates/postgresql.conf template.j2:
# ...
buffer: "{{ansible_facts['memtotal_mb'] / 10 }}"
# ...
Total:
In this lesson, we learned how to use facts in Ansible to get information about the target machine.
Facts are static information about the server, such as the operating system, processor, memory, and other settings.
The facts are stored in the ansible_facts variable and can be used in playbooks to dynamically configure tasks.
In this example, we used the fact ansible_facts['memtotal_mb'] to determine the buffer size in the PostgreSQL configuration file.